SemiProject: member(dropdown), JSP: EL language

+++ Members/WebContent/memberDelete.jsp
@@ -8,9 +8,20 @@
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <jsp:useBean id="memDAO" class="com.jweb.member.MemberDAO" scope="application"/>
 <%
-	String memId = request.getParameter("memberId");
-	memDAO.deleteMember(memId);
-	response.sendRedirect("memberList.jsp");
+	
+	String sessionId = null;
+	if(session.getAttribute("sessionId") != null){
+		sessionId = (String) session.getAttribute("sessionId");
+		String memId = request.getParameter("memberId");
+		memDAO.deleteMember(memId);
+		response.sendRedirect("logout.jsp?msg=0");
+	} else {
+		out.println("<script>");
+		out.println("alert('로그인을 해주세요.')");
+		out.println("location.href='loginForm.jsp'");
+		out.println("</script>");
+	}
+	
 %>
 </head>
 <body>
